Clean up LinodeResize_CMR
